Refactor cart count functionality and update header/footer rendering
